+++ Omega/res/raw/changelog.md
@@ -1,5 +1,6 @@
 ### v0.4.0 (Build 30)
-* **Add hide apps preference**
+* **Add: Hide apps preference**
+* **Add: Icon Shape preference**
 
 ### v0.3.0 (Build 20)
 * **Add gestures preference**